backend/requirements.txt
# ScheduleAI Backend — Python Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# AI / LangGraph
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# MongoDB (async)
motor>=3.5.0
pymongo>=4.9.0
bcrypt>=4.0.0



# HTTP
httpx>=0.27.0
